fix(sonar): Cast one of the operands to long L411

+++ src/aggregation/aggregationcontroller_api/src/main/java/se/tink/backend/aggregation/aggregationcontroller/AggregationControllerAggregationClientImpl.java
@@ -57,7 +57,7 @@ public class AggregationControllerAggregationClientImpl
     private static final ImmutableSet<String> IDENTITY_AGGREGATOR_ENABLED_ENVIRONMENTS =
             ImmutableSet.of("oxford-staging", "oxford-production");
     private static final int MAXIMUM_RETRY_ATTEMPT = 3;
-    private static final int WAITING_TIME_FOR_NEW_ATTEMPT_IN_MILLISECONDS = 2000;
+    private static final long WAITING_TIME_FOR_NEW_ATTEMPT_IN_MILLISECONDS = 2000;
     private static final ImmutableSet<Integer> ERROR_CODES_FOR_RETRY =
             ImmutableSet.of(502, 503, 504);
